requirements.txt
beautifulsoup4==4.10.0
bs4==0.0.1
cloudscraper==1.2.58
html5lib==1.1
lxml==4.7.1
mysql-connector-python==8.0.27
mysqlclient==2.1.0
requests==2.26.0
webdriver-manager==3.5.2
